dev-support: move over the Fifo
eventually basically everything from misc/lc3-tm4c will migrate over to:
  - this crate (for stuff that's not specific to the t4mc)
  - or the lc3-tm4c crate that lives outside this repo (for stuff that *is* specific to the tm4c)
