+++ device-support/src/util/mod.rs
@@ -1,0 +1,5 @@
+//! Miscellaneous helper things. (TODO)
+
+mod fifo;
+
+pub use fifo::Fifo;
